+++ CHANGELOG.md
@@ -4,6 +4,14 @@ Ce projet suit [Semantic Versioning](https://semver.org/lang/fr/).
 
 ## [Non publié]
 
+## [1.2.0] - 2026-08-21
+
+### Ajouté
+
+- option de barre des menus pour activer ou désactiver le lancement automatique
+  de Capote à l’ouverture de session, avec lecture de l’état réel de macOS et
+  accès aux Réglages Système lorsqu’une approbation est requise.
+
 ## [1.1.0] - 2026-08-21
 
 ### Ajouté